Split infoenergia.py in two (enviament & lot). Add field autoconsum & tarifa. Minor fixes
