+++ som_infoenergia/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
-import infoenergia
+import som_infoenergia_lot
+import som_infoenergia_enviament
 import wizard
 import giscedata_polissa
